Add files via upload
You are a data analyst and your client has a large ecommerce company in India (let’s call it X). 
X gets a thousand orders via their website on a daily basis and they have to deliver them as fast 
as they can. For delivering the goods ordered by the customers, X has tied up with multiple 
courier companies in India as delivery partners who charge them some amount per delivery.
The charges are dependent upon two factors:
● Weight of the product
● Distance between the warehouse (pickup location) and customer’s delivery address 
(destination location)
On an average, the delivery charges are Rs. 100 per shipment. So if X ships 1,00,000 orders 
per month, they have to pay approximately Rs. 1 crore to the courier companies on a monthly 
basis as charges.
As the amount that X has to pay to the courier companies is very high, they want to verify if the 
charges levied by their Delivery partners per Order are correct.